Removed need for mongod to be in PATH (made path to mongod configurable).

+++ NoRM.Tests/MongoFixture.cs
@@ -1,5 +1,8 @@
 using System;
 using System.Diagnostics;
+using System.Configuration;
+
+using Xunit;
 
 namespace Norm.Tests
 {
@@ -15,6 +18,9 @@ namespace Norm.Tests
         protected virtual string Arguments
         {
             get { return string.Empty; }
+        }
+        protected virtual string MongodPath {
+            get { return ConfigurationManager.AppSettings["mongodPath"]; }
         }
 
         protected string ConnectionString()
@@ -48,16 +54,40 @@ namespace Norm.Tests
               {
                   StartInfo = new ProcessStartInfo
                           {
-                              FileName = "mongod",
+                              UseShellExecute = false,
+                              FileName = this.MongodPath,
                               Arguments = string.Format("--port {0} --dbpath {1} --smallfiles {2}",
                                 this.Port, this.DataPath, this.Arguments),
-                          },
+                              RedirectStandardError = true,
+                              RedirectStandardOutput = true
+                          }
               };
-            _process.Start();
+            _process.Exited += _process_Exited;
+            _process.Start();            
+        }
+        
+        private void _process_Exited(object sender, EventArgs e) {
+            if (_process.ExitCode == 0)
+            {
+                return;
+            }
+
+            throw new InvalidOperationException(string.Format(
+                "Mongod process exited unexpectedly.{0}Output:{0}{1}.{0}Errors:{0}{2}.",
+                Environment.NewLine,
+                _process.StandardOutput.ReadToEnd(),
+                _process.StandardError.ReadToEnd()
+            ));
         }
 
         public void Dispose()
         {
+            if (_process.HasExited)
+            {
+                return;
+            }
+
+            _process.Exited -= _process_Exited;
             _process.Kill();
         }
     }
